fix FLAGS + norm

+++ libft/ft_substr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_substr.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ffiliz <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: mgoudin <mgoudin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/21 00:22:24 by ffiliz            #+#    #+#             */
-/*   Updated: 2022/06/21 00:22:31 by ffiliz           ###   ########.fr       */
+/*   Updated: 2022/06/21 16:32:47 by mgoudin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,9 +18,9 @@ char	*ft_substr(char *s, unsigned int start, size_t len)
 	char	*str;
 
 	i = 0;
-	if (!s || len <= 0 || start >= ft_strlen(s))
+	if (!s || len <= 0 || start >= (unsigned int)ft_strlen(s))
 		return (ft_strdup(""));
-	if (len > ft_strlen(s))
+	if (len > (size_t)ft_strlen(s))
 		return (ft_strdup((char *)s + start));
 	str = (char *)malloc(sizeof(char) * len + 1);
 	if (!str)
